chore: modularize server
